requirements.txt
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.1.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-engineio==3.0.0
python-socketio==2.1.0
RPi.GPIO==0.6.5
six==1.11.0
Werkzeug==0.14.1
